GT-2104 Move logic to Extension
